+++ src/core/ext/transport/chttp2/transport/hpack_parser_table.cc
@@ -229,7 +229,8 @@ HPackTable::Memento MakeMemento(size_t i) {
   auto sm = kStaticTable[i];
   return grpc_metadata_batch::Parse(
       sm.key, Slice::FromStaticString(sm.value),
-      strlen(sm.key) + strlen(sm.value) + hpack_constants::kEntryOverhead,
+      static_cast<uint32_t>(strlen(sm.key) + strlen(sm.value) +
+                            hpack_constants::kEntryOverhead),
       [](absl::string_view, const Slice&) {
         abort();  // not expecting to see this
       });
